% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg-methods}
\alias{linreg-methods}
\alias{print.linreg}
\alias{plot.linreg}
\alias{residuals.linreg}
\alias{pred}
\alias{pred.linreg}
\alias{coef.linreg}
\alias{summary.linreg}
\title{Methods for linreg objects}
\usage{
\method{print}{linreg}(x, ...)

\method{plot}{linreg}(x, ...)

\method{residuals}{linreg}(object, ...)

pred(object, ...)

\method{pred}{linreg}(object, ...)

\method{coef}{linreg}(object, ...)

\method{summary}{linreg}(object, ...)
}
\arguments{
\item{x, object}{An object of class \code{linreg}.}

\item{...}{Further arguments (currently unused).}
}
\value{
\code{print}, \code{summary} and \code{plot} are called for their
side effects. \code{resid}/\code{residuals} returns the residuals,
\code{pred} the fitted values and \code{coef} the named coefficients.
}
\description{
Print, summarise, plot and extract results from a fitted \code{linreg} model.
}
