% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object (e.g., \code{Petal.Length ~ Species}).}

\item{data}{A data frame containing the variables.}
}
\value{
An object of class \code{linreg}.
}
\description{
Fits a linear regression model using linear algebra.
}
