\name{linregpackage-package}
\alias{linregpackage-package}
\alias{linregpackage}
\docType{package}
\title{
\packageTitle{linregpackage}
}
\description{
\packageDescription{linregpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{linregpackage}
\packageIndices{linregpackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{linregpackage}

Maintainer: \packageMaintainer{linregpackage}
}
\references{
#~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{

}
