% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Diagnostic Plots for a linreg Model}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments, currently not used.}
}
\value{
A list containing two ggplot objects.
}
\description{
Creates diagnostic plots for a fitted linreg model.
}
\details{
The method produces two plots:
\itemize{
  \item Residuals vs Fitted
  \item Scale-Location
}
}
