% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.linreg}
\title{Extract Predicted Values from a linreg Model}
\usage{
pred(object, ...)

\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments, currently not used.}
}
\value{
A numeric vector containing the predicted values.
}
\description{
Returns the predicted values from a fitted linreg model.
}
