% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.linreg}
\alias{predict.linreg}
\title{A method that returns the predicted values y.}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{x}{linreg object}
}
\value{
vector of predictions.
}
\description{
A method that returns the predicted values y.
}

