% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg Model}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}.}

\item{...}{Additional arguments, currently not used.}
}
\value{
The linreg object, returned invisibly.
}
\description{
Prints the model call and estimated regression coefficients
from a fitted linreg model.
}
