% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract Residuals from a linreg Model}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}.}

\item{...}{Additional arguments, currently not used.}
}
\value{
A numeric vector containing the residuals.
}
\description{
Returns the residuals from a fitted linreg model.
}
