% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{A function for ridge regression.}
\usage{
ridgereg(formula, data)
}
\arguments{
\item{formula}{A formula implies the model (e.g. y ~ x)}

\item{data}{data frame, dataset attached to the algorithm}
}
\value{
ridgereg object.
}
\description{
A function for ridge regression.
}
\examples{
will come soon
}

