% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_colour_liu}
\alias{scale_colour_liu}
\title{Linkoping University Colour Scale}
\usage{
scale_colour_liu(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{ggplot2::scale_colour_manual()}.}
}
\value{
A ggplot2 discrete colour scale.
}
\description{
Applies colours from the Linkoping University graphical profile
to discrete colour aesthetics.
}
