% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_liu}
\alias{scale_fill_liu}
\title{Linkoping University Fill Scale}
\usage{
scale_fill_liu(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{ggplot2::scale_fill_manual()}.}
}
\value{
A ggplot2 discrete fill scale.
}
\description{
Applies colours from the Linkoping University graphical profile
to discrete fill aesthetics.
}
