% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_liu}
\alias{theme_liu}
\title{Linkoping University ggplot2 Theme}
\usage{
theme_liu(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Creates a ggplot2 theme based on the graphical profile
of Linkoping University.
}
