% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/airports.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{A function that creates a plot showing mean delay of flights.}
\usage{
visualize_airport_delays()
}
\arguments{
\item{no}{arguments}
}
\description{
A function that creates a plot showing mean delay of flights.
}
\examples{
#library(ggplot2)
#library(dplyr)
#library(nycflights13)
}

