% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{The plot-method that creates two plots.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\description{
The plot-method that creates two plots.
}

