% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{A method that prints a summary of the linreg-function.}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{x}{linreg object}
}
\value{
linreg_summary object, can be printed.
}
\description{
A method that prints a summary of the linreg-function.
}

